+++ dataiter/io.py
@@ -54,7 +54,11 @@ def read_json(path, *, encoding="utf-8", keys=[], types={}, **kwargs):
 def read_npz(path, *, allow_pickle=True):
     return DataFrame.read_npz(path, allow_pickle=allow_pickle)
 
+def read_parquet(path, *, columns=[], strings_as_object=inf, dtypes={}):
+    return DataFrame.read_parquet(path, columns=[], strings_as_object=inf, dtypes={})
+
 read_csv.__doc__ = util.format_alias_doc(read_csv, DataFrame.read_csv)
 read_geojson.__doc__ = util.format_alias_doc(read_geojson, GeoJSON.read)
 read_json.__doc__ = util.format_alias_doc(read_json, ListOfDicts.read_json)
 read_npz.__doc__ = util.format_alias_doc(read_npz, DataFrame.read_npz)
+read_parquet.__doc__ = util.format_alias_doc(read_parquet, DataFrame.read_parquet)